(PC-12635) feat(DateInput): update test
